doc: remove misleading documentation on pack names
The index-pack documentation explicitly states that the pack
name is derived from the sorted list of object names, but
that clearly isn't true. I can't seem to be able to figure
out if this was ever changed though.

Be less explicit in the docs as to what the exact output is,
and just say that it's whatever goes into the pack name.

Really it seems to be the sha1 of the entire file, without
the checksum footer.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>